# api/swagger/swagger.json
  {
    "swagger": "2.0",
    "info": {
      "description": "creating Sup api documantation using swagger ",
      "version": "1.0.0",
      "title": "Sup Api"
    },
    "host": " 192.168.10.20:3000",
    "basePath": "/kitchens",
...
    "schemes": [
      "https",
      "http"
    ],
    "paths": {
...
      "/categories/getcategoriesbyid/{categoryId}": {
        "get": {
          "tags": [
            "Categories"
          ],
          "summary": "Find Category by ID",
          "description": "Returns a single Category",
          "operationId": "getCategoriesById",
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "name": "categoryId",
              "in": "req.body",
              "description": "ID of Category to return",
              "required": true,
              "type": "mongoose.Types.ObjectId()",
              "format": "mongooseObject"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation"
             
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "kitchen not found"
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
        } 
      },
      "/categories/getcategories": {
        "get": {
          "tags": [
            "Categories"
          ],
          "summary": "fetch All Categories ",
          "description": "Returns Categories Collection",          
          "produces": [
            "application/xml",
            "application/json"
          ],          
          "responses": {
            "200": {
              "description": "successful operation"
             
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "kitchen not found"
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
        }
      },
      "/categories/addcategories": {
        "post": {
          "tags": [
            "Categories"
          ],
          "summary": "Add a new Category to the collection",
          "description": "",
          "operationId": "addCategory",
          "consumes": [
            "application/json",
            "application/xml"
          ],
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Category object that needs to be added to the Collection",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Categories"
              }
            }
          ],
          "responses": {
            "405": {
              "description": "Invalid input"
            }
          },
          "security": [
            {
              "petstore_auth": [
                "write:category",
                "read:category"
              ]
            }
          ]
        }       
      },
...
      "/users/deletecategories/{categoryId}": {
        "delete": {
          "tags": [
            "Categories"
          ],
          "summary": "Delete a Category",
          "description": "",
          "operationId": "deleteCategory",
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "name": "api_key",
              "in": "req.params",
              "required": false,
              "type": "string"
            },
            {
              "name": "categoryId",
              "in": "req.params",
              "description": "category id to delete",
              "required": true,
              "type": "mongoose.Types.ObjectId",
              "format": "mongoose.schema.ObjectId"
            }
          ],
          "responses": {
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Pet not found"
            }
          },
          "security": [
            {
              "petstore_auth": [
                "write:Category",
                "read:Category"
              ]
            }
          ]
        }
      }
    },   
    "securityDefinitions": {
      "petstore_auth": {
        "type": "oauth2",
        "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
        "flow": "implicit"
        
      },
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    },
    "definitions": {     
...
      "Categories": {
        "type": "object",
        "properties": {
          "id": {
            "type": "mongoose.Types.ObjectId",
            "format": "mongoose.schema.objectId"
          },
          "companyName": {
            "type": "string"
          },
          "trnNumber": {
            "type": "string"
          }, 
          "supplyItem": {
            "type": "string"           
          },         
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "userType": {
            "type": "string"           
           
          },
          "createdDate": {
            "type": "DateTime"           
           
          },
          "updatedDate": {
            "type": "DateTime"           
           
          },
          "user": {
            "type": "mongoose.Types.ObjectId"           
           
          }
        },
        "xml": {
          "name": "supplier"
        }
      }
    }
  }